fix: locationService 따로 뺐음
App 시작점에서만 권한 요청하도록 수정했습니다 !
